Move all test constants to single file
