+++ app/src/androidTest/java/com/github/fribourgsdp/radio/activities/AddPlaylistActivityWithoutConnectionTest.kt
@@ -13,6 +13,7 @@ import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.data.User
 import com.github.fribourgsdp.radio.data.view.UserProfileActivity
 import com.github.fribourgsdp.radio.mockimplementations.MockAddPlaylistActivityNoConnection
+import com.github.fribourgsdp.radio.utils.*
 import com.google.android.gms.tasks.Tasks
 import org.hamcrest.Matchers
 import org.junit.After
@@ -52,7 +53,7 @@ class AddPlaylistActivityWithoutConnectionTest {
 
         val user = Tasks.await(User.loadOrDefault(ctx))
         assert(user.getPlaylists().any { p -> p.name == testPlaylist1 })
-        user.getPlaylists().filter { p -> p.name == testPlaylist1}.forEach{ p ->
+        user.getPlaylists().filter { p -> p.name == testPlaylist1 }.forEach{ p ->
             run {
                 assert(p.getSongs().any { s -> s.name == testSong1 && s.artist == testArtist })
                 assert(p.getSongs().any { s -> s.name == testSong2 && s.artist == testArtist })
