+++ app/src/androidTest/java/com/github/fribourgsdp/radio/activities/GameSettingsActivityTest.kt
@@ -22,6 +22,8 @@ import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.data.Playlist
 import com.github.fribourgsdp.radio.game.prep.*
 import com.github.fribourgsdp.radio.mockimplementations.MockGameSettingsActivity
+import com.github.fribourgsdp.radio.utils.packageName
+import com.github.fribourgsdp.radio.utils.testPlaylist1
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import org.hamcrest.Matcher
