+++ app/src/androidTest/java/com/github/fribourgsdp/radio/activities/GoogleSignInActivityTest.kt
@@ -15,6 +15,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.data.view.UserProfileActivity
 import com.github.fribourgsdp.radio.deprecated.GoogleSignInActivity
+import com.github.fribourgsdp.radio.utils.packageName
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.auth.*
 import org.hamcrest.Matchers
