+++ app/src/androidTest/java/com/github/fribourgsdp/radio/activities/VoiceOverIPActivityTest.kt
@@ -10,6 +10,7 @@ import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.rule.GrantPermissionRule
 import com.github.fribourgsdp.radio.MainActivity
 import com.github.fribourgsdp.radio.deprecated.VoiceOverIPActivity
+import com.github.fribourgsdp.radio.utils.packageName
 import org.hamcrest.Matchers
 import org.junit.After
 import org.junit.Before
