+++ app/src/androidTest/java/com/github/fribourgsdp/radio/external/VoiceOverIpEngineDecoratorTest.kt
@@ -8,7 +8,7 @@ import androidx.test.espresso.Espresso
 import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.matcher.IntentMatchers
 import com.github.fribourgsdp.radio.MainActivity
-import com.github.fribourgsdp.radio.activities.packageName
+import com.github.fribourgsdp.radio.utils.packageName
 import com.github.fribourgsdp.radio.deprecated.VoiceOverIPActivity
 import com.github.fribourgsdp.radio.voip.VoiceIpEngineDecorator
 import io.agora.rtc.IRtcEngineEventHandler
