+++ app/src/androidTest/java/com/github/fribourgsdp/radio/ui/PlaylistRecyclerViewTest.kt
@@ -6,17 +6,15 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.assertion.ViewAssertions
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.contrib.RecyclerViewActions
 import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.github.fribourgsdp.radio.R
 import com.github.fribourgsdp.radio.mockimplementations.MockUserProfileActivity
-import com.github.fribourgsdp.radio.mockimplementations.playListName
-import com.github.fribourgsdp.radio.mockimplementations.songName
+import com.github.fribourgsdp.radio.utils.playListName
+import com.github.fribourgsdp.radio.utils.songName
 import com.github.fribourgsdp.radio.util.ViewHolder
 import com.google.android.gms.tasks.Tasks
 import com.google.firebase.auth.FirebaseAuth
