+++ app/src/androidTest/java/com/github/fribourgsdp/radio/utils/TestConstants.kt
@@ -1,0 +1,37 @@
+package com.github.fribourgsdp.radio.utils
+
+const val testLyrics6 = "If you feel, little chance, make a stance\n" +
+        "Looking for, better days, let me say\n" +
+        "Something's wrong, when you can't, let me go\n" +
+        "For to long, long, long...\n" +
+        "\n" +
+        "Momentum owns you\n" +
+        "Controlling her too"
+const val longLyrics =
+    " Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus et lobortis elit, non rutrum sem. Proin at sapien ac justo molestie scelerisque non ac enim. Cras fermentum, massa ac maximus suscipit, ex diam dapibus nulla, in mattis enim ligula ut eros. Etiam nec libero nibh. Phasellus scelerisque purus eu orci congue cursus ut tempus mi. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Aliquam luctus consequat tellus, vitae interdum felis molestie vitae. Donec dapibus tellus vel sapien laoreet tempor. Suspendisse sodales in magna eu vehicula. Mauris faucibus risus a leo vestibulum, ac vestibulum elit sodales. "
+const val packageName = "com.github.fribourgsdp.radio"
+const val testArtist = "Sardou"
+const val testSong1 = "Rouge"
+const val testSong2 = "En chantant"
+const val testSong3 = "Le France"
+const val testSong5 = "salut"
+const val testSong6 = "Momentum"
+const val testArtist6 = "Truckfighters"
+const val testPlaylist1 = "Sardou playlist"
+const val testPlaylist2 = "Stoner Playlist"
+const val userName = "test"
+const val playListName = "testTitle"
+const val songName = "TestSongName"
+const val userId = "testId"
+const val artistName = "artist"
+const val playlistName = "testPlaylist"
+const val testPlaylistId = "TEST_PLAYLIST"
+const val onlineUserId = "onlineUserTestId"
+const val onlineUserName = "onlineUserTest"
+const val testSinger1 = "TheBestSinger"
+const val testSinger2 = "TheOverlord"
+const val testSinger3 = "TheBreadMaster"
+const val testLyrics4 =
+    "A long, long time ago\nI can still remember\nHow that music used to make me smile..."
+const val testSong4 = "American Pie"
+const val testPlaylist4 = "70s hits"
